+++ pkg/exams/R/exams2nops.R
@@ -3,7 +3,7 @@ exams2nops <- function(file, n = 1L, nsamp = NULL, dir = ".", name = NULL,
   institution = "R University", logo = "Rlogo.png", date = Sys.Date(), 
   replacement = FALSE, intro = NULL, blank = NULL, duplex = TRUE, pages = NULL,
   usepackage = NULL, header = NULL, encoding = "UTF-8", startid = 1L, points = NULL,
-  showpoints = FALSE, samepage = FALSE, twocolumn = FALSE, reglength = 7L, seed = NULL, ...)
+  showpoints = FALSE, samepage = FALSE, newpage = FALSE, twocolumn = FALSE, reglength = 7L, seed = NULL, ...)
 {
   ## handle matrix specification of file
   if(is.matrix(file)) {
@@ -131,7 +131,7 @@ exams2nops <- function(file, n = 1L, nsamp = NULL, dir = ".", name = NULL,
     replacement = replacement, intro = intro, blank = blank,
     duplex = duplex, pages = pages, file = template,
     nchoice = nchoice,
-    encoding = encoding, samepage = samepage, twocolumn = twocolumn, reglength = reglength)
+    encoding = encoding, samepage = samepage, newpage = newpage, twocolumn = twocolumn, reglength = reglength)
 
   ## if points should be shown generate a custom transformer
   transform <- if(showpoints) {
@@ -171,7 +171,7 @@ exams2nops <- function(file, n = 1L, nsamp = NULL, dir = ".", name = NULL,
 
 make_nops_template <- function(n, replacement = FALSE, intro = NULL, blank = NULL,
   duplex = TRUE, pages = NULL, file = NULL, nchoice = 5, encoding = "UTF-8",
-  samepage = FALSE, twocolumn = FALSE, reglength = 7L)
+  samepage = FALSE, newpage = FALSE, twocolumn = FALSE, reglength = 7L)
 {
 
 page1 <- make_nops_page(n, nchoice = nchoice, reglength = reglength)
@@ -334,14 +334,20 @@ sprintf("\\reg%s%s", c("seven", "eight", "nine", "ten"), tolower(0L:3L == addreg
 \\fi
 
 %% for exams2pdf
-\\newenvironment{question}{\\item}{}
 \\newenvironment{solution}{\\comment}{\\endcomment}",
+
+if(newpage) {
+  "\\newenvironment{question}{\\item}{\\newpage}"
+} else {
+  "\\newenvironment{question}{\\item}{}"
+},
 
 if(samepage) {
   "\\newenvironment{answerlist}{\\renewcommand{\\labelenumii}{(\\alph{enumii})}\\begin{samepage}\\begin{enumerate}}{\\end{enumerate}\\end{samepage}}"
 } else {
   "\\newenvironment{answerlist}{\\renewcommand{\\labelenumii}{(\\alph{enumii})}\\begin{enumerate}}{\\end{enumerate}}"
 },
+
 "
 %% additional header commands
 \\makeatletter